+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Fixed
 
 - `ui/Marquee` to marquee when necessary after a locale change
+
 ## [2.2.8] - 2018-12-06
 
 ### Fixed